Further cleaned up the Tetris API.
Finished splitting it out into files. Also, enabled and resolved all of
the -W warnings.

Resolved a notable bug in the Grenade.Exts UpdateRaw implementation,
which ensured only the first layer could be optimized.

Also, optimized the benchmark advance function to no longer simulate
HardDrop when searching for moves. This yielded a 10x speed up in the
benchmark.

Signed-off-by: Joshua Brot <jbrot@umich.edu>